+++ pages/stories/[id].js
@@ -1,19 +1,9 @@
 import { useRouter } from "next/router";
 import StoryDetailCard from "@/components/Story";
-import { useState, useEffect } from "react";
 
-function StoryDetailPage() {
-  const [stories, setStories] = useState([]);
-
+function StoryDetailPage({ stories }) {
   const router = useRouter();
   const { id } = router.query;
-  useEffect(() => {
-    const storedStories = JSON.parse(localStorage.getItem("stories"));
-
-    if (storedStories && Array.isArray(storedStories)) {
-      setStories(storedStories);
-    }
-  }, []);
 
   const story = stories.find((story) => story.id === parseInt(id));
